Move Timestamp marshalling to internal/api, use gqlgen Int64 parser
