+++ gqlgen.yml
@@ -23,10 +23,10 @@ autobind:
 
 models:
   # Scalars
-  Timestamp:
-    model: github.com/stashapp/stash/pkg/models.Timestamp
   Int64:
-    model: github.com/stashapp/stash/pkg/models.Int64
+    model: github.com/99designs/gqlgen/graphql.Int64
+  Timestamp:
+    model: github.com/stashapp/stash/internal/api.Timestamp
   # define to force resolvers
   Image:
     model: github.com/stashapp/stash/pkg/models.Image
