+++ internal/api/models.go
@@ -1,4 +1,4 @@
-package models
+package api
 
 import (
 	"errors"
