+++ internal/api/models_test.go
@@ -1,4 +1,4 @@
-package models
+package api
 
 import (
 	"bytes"
